--- myApp/routes/users.js
@@ -1,11 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* GET users listing. */
-router.get('/', function(req, res, next) {
-  res.io.emit("socketToMe", "users");
-  res.send('respond with a resource.');
-});
-
-module.exports = router;
-
