--- myApp/views/index.jade
@@ -1,27 +1,0 @@
-extends layout
-block content  
-  
-  h1= title
-
-  p Welcome to #{title}
-
-  script(src="/socket.io/socket.io.js")
-
-  script.
-
-    var socket = io('//'+document.location.hostname+':'+document.location.port);
-
-    socket.on('socketToMe', function (data) {
-
-      console.log(data);
-
-    });
-
-
-//- extends layout
-
-//- block content
-//-   h1= title
-//-   p Welcome to #{title}
-
-
