--- myApp/views/layout.jade
@@ -1,12 +1,0 @@
-doctype html
-html
-  head
-    title= title
-    link(rel='stylesheet', href='/stylesheets/style.css')
-  body
-    block content
-
-
-
-
-
