+++ routes/index.js
@@ -6,4 +6,4 @@ router.get('/', function(req, res, next) {
   res.render('index', { title: 'Express' });
 });
 
-module.exports = router; 
+module.exports = router;
