+++ routes/users.js
@@ -8,3 +8,4 @@ router.get('/', function(req, res, next) {
 });
 
 module.exports = router;
+
