+++ views/canvaschart.jade
