+++ views/chart.jade
