+++ views/slider.jade
